pyproject.toml
[tool.poetry]
name = "fastapi-todo-app"
version = "1.0.0"
description = "A robust RESTful API built with FastAPI for managing todos with user authentication and authorization"
authors = ["Your Name <nhan2151@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [
    { include = "*.py" },
    { include = "routers/*.py" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Framework :: FastAPI",
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
gunicorn = "^23.0.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
bcrypt = "^4.2.1"
python-multipart = "^0.0.19"
SQLAlchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
pydantic = "^2.10.3"
pydantic-settings = "^2.0.0"
starlette = "^0.41.3"
celery = "^5.4.0"
pytest-asyncio = "^0.23.0"
pytest = "^8.3.4"
Jinja2 = "^3.1.4"
prometheus-client = "^0.19.0"
prometheus-fastapi-instrumentator = "^6.1.0"
opentelemetry-api = "1.21.0"
opentelemetry-sdk = "1.21.0"
opentelemetry-instrumentation-fastapi = "0.42b0"
opentelemetry-exporter-otlp = "1.21.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^4.1.0"
black = "^24.10.0"
flake8 = "^7.0.0"
mypy = "^1.9.0"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
^/docs
'''

[tool.ruff]
# Enable pycodestyle ('E'), pyflakes ('F'), and isort ('I') codes
select = ["E", "F", "I"]
ignore = ["E501"]  # Line length is handled by black

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "I"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

[tool.ruff.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=./ --cov-report=term-missing --cov-report=xml"
pythonpath = "."
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
]

[tool.coverage.run]
source = ["."]
omit = [
    "tests/*",
    "venv/*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "pass",
    "raise ImportError",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
show_error_codes = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["sqlalchemy.*", "passlib.*", "jose.*", "pytest.*"]
ignore_missing_imports = true

[tool.bandit]
exclude_dirs = ["tests", "venv", ".git", "__pycache__"]
skips = ["B101"]  # Skip assert statements
targets = ["."]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88 